refactor(session): move levels and mission from studylog to session
