+++ backend/src/acceptanceTest/java/wooteco/prolog/fixtures/MissionAcceptanceFixture.java
@@ -1,6 +1,6 @@
 package wooteco.prolog.fixtures;
 
-import wooteco.prolog.studylog.application.dto.MissionRequest;
+import wooteco.prolog.session.application.dto.MissionRequest;
 
 public class MissionAcceptanceFixture {
