+++ backend/src/acceptanceTest/java/wooteco/prolog/steps/MissionStepDefinitions.java
@@ -10,8 +10,8 @@ import org.springframework.http.HttpStatus;
 import wooteco.prolog.AcceptanceSteps;
 import wooteco.prolog.fixtures.LevelAcceptanceFixture;
 import wooteco.prolog.fixtures.MissionAcceptanceFixture;
-import wooteco.prolog.studylog.application.dto.MissionRequest;
-import wooteco.prolog.studylog.application.dto.MissionResponse;
+import wooteco.prolog.session.application.dto.MissionRequest;
+import wooteco.prolog.session.application.dto.MissionResponse;
 
 public class MissionStepDefinitions extends AcceptanceSteps {
