+++ backend/src/main/java/wooteco/prolog/session/application/MissionService.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.studylog.application;
+package wooteco.prolog.session.application;
 
 import static java.util.stream.Collectors.toList;
 
@@ -6,11 +6,11 @@ import java.util.List;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.prolog.studylog.application.dto.MissionRequest;
-import wooteco.prolog.studylog.application.dto.MissionResponse;
-import wooteco.prolog.studylog.domain.Level;
-import wooteco.prolog.studylog.domain.Mission;
-import wooteco.prolog.studylog.domain.repository.MissionRepository;
+import wooteco.prolog.session.application.dto.MissionRequest;
+import wooteco.prolog.session.application.dto.MissionResponse;
+import wooteco.prolog.session.domain.Level;
+import wooteco.prolog.session.domain.Mission;
+import wooteco.prolog.session.domain.repository.MissionRepository;
 import wooteco.prolog.studylog.exception.DuplicateMissionException;
 import wooteco.prolog.studylog.exception.MissionNotFoundException;
