+++ backend/src/main/java/wooteco/prolog/session/domain/repository/LevelRepository.java
@@ -1,8 +1,8 @@
-package wooteco.prolog.studylog.domain.repository;
+package wooteco.prolog.session.domain.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import wooteco.prolog.studylog.domain.Level;
+import wooteco.prolog.session.domain.Level;
 
 public interface LevelRepository extends JpaRepository<Level, Long> {
