+++ backend/src/main/java/wooteco/prolog/session/domain/repository/MissionRepository.java
@@ -1,8 +1,8 @@
-package wooteco.prolog.studylog.domain.repository;
+package wooteco.prolog.session.domain.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import wooteco.prolog.studylog.domain.Mission;
+import wooteco.prolog.session.domain.Mission;
 
 public interface MissionRepository extends JpaRepository<Mission, Long> {
