+++ backend/src/main/java/wooteco/prolog/session/ui/LevelController.java
@@ -1,4 +1,4 @@
-package wooteco.prolog.studylog.ui;
+package wooteco.prolog.session.ui;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.prolog.studylog.application.LevelService;
-import wooteco.prolog.studylog.application.dto.LevelRequest;
-import wooteco.prolog.studylog.application.dto.LevelResponse;
+import wooteco.prolog.session.application.LevelService;
+import wooteco.prolog.session.application.dto.LevelRequest;
+import wooteco.prolog.session.application.dto.LevelResponse;
 
 @RestController
 @RequestMapping("/levels")
@@ -19,15 +19,15 @@ public class LevelController {
 
     private final LevelService levelService;
 
-    @GetMapping
-    public ResponseEntity<List<LevelResponse>> show() {
-        List<LevelResponse> responses = levelService.findAll();
-        return ResponseEntity.ok(responses);
-    }
-
     @PostMapping
     public ResponseEntity<LevelResponse> create(@RequestBody LevelRequest levelRequest) {
         LevelResponse levelResponse = levelService.create(levelRequest);
         return ResponseEntity.ok(levelResponse);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<LevelResponse>> show() {
+        List<LevelResponse> responses = levelService.findAll();
+        return ResponseEntity.ok(responses);
     }
 }
