+++ backend/src/main/java/wooteco/prolog/studylog/application/dto/FilterResponse.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import wooteco.prolog.member.application.dto.MemberResponse;
+import wooteco.prolog.session.application.dto.LevelResponse;
+import wooteco.prolog.session.application.dto.MissionResponse;
 
 @NoArgsConstructor
 @AllArgsConstructor
