+++ backend/src/test/java/wooteco/prolog/studylog/domain/StudylogTagsTest.java
@@ -11,6 +11,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.prolog.member.domain.Member;
 import wooteco.prolog.member.domain.Role;
+import wooteco.prolog.session.domain.Level;
+import wooteco.prolog.session.domain.Mission;
 
 public class StudylogTagsTest {
