+++ backend/src/test/java/wooteco/prolog/studylog/domain/StudylogTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.prolog.member.domain.Member;
 import wooteco.prolog.member.domain.Role;
+import wooteco.prolog.session.domain.Level;
+import wooteco.prolog.session.domain.Mission;
 
 class StudylogTest {
