+++ backend/src/test/java/wooteco/prolog/studylog/domain/repository/ReportRepositoryTest.java
@@ -23,8 +23,10 @@ import wooteco.prolog.member.domain.repository.MemberRepository;
 import wooteco.prolog.report.application.dto.report.ReportAssembler;
 import wooteco.prolog.report.domain.ablity.repository.AbilityRepository;
 import wooteco.prolog.report.domain.report.repository.ReportRepository;
-import wooteco.prolog.studylog.domain.Level;
-import wooteco.prolog.studylog.domain.Mission;
+import wooteco.prolog.session.domain.Level;
+import wooteco.prolog.session.domain.Mission;
+import wooteco.prolog.session.domain.repository.LevelRepository;
+import wooteco.prolog.session.domain.repository.MissionRepository;
 import wooteco.prolog.studylog.domain.Studylog;
 import wooteco.prolog.report.domain.ablity.Ability;
 import wooteco.prolog.report.domain.report.Report;
