+++ backend/src/test/java/wooteco/prolog/studylog/studylog/domain/repository/StudylogReadRepositoryTest.java
@@ -9,8 +9,10 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import wooteco.prolog.member.domain.Member;
 import wooteco.prolog.member.domain.Role;
 import wooteco.prolog.member.domain.repository.MemberRepository;
-import wooteco.prolog.studylog.domain.Level;
-import wooteco.prolog.studylog.domain.Mission;
+import wooteco.prolog.session.domain.Level;
+import wooteco.prolog.session.domain.Mission;
+import wooteco.prolog.session.domain.repository.LevelRepository;
+import wooteco.prolog.session.domain.repository.MissionRepository;
 import wooteco.prolog.studylog.domain.Studylog;
 import wooteco.prolog.studylog.domain.StudylogRead;
 import wooteco.prolog.studylog.domain.repository.*;
